% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD_s3method.R
\name{create_clustvarselect}
\alias{create_clustvarselect}
\title{Create ClustVarSelect Object}
\usage{
create_clustvarselect(scores, k_guess, prop_algo, class_algo, params = list())
}
\arguments{
\item{scores}{Numeric vector of feature importance scores}

\item{k_guess}{Number of clusters used}

\item{prop_algo}{Clustering algorithm used}

\item{class_algo}{Classification algorithm used}

\item{params}{List of additional parameters used}
}
\value{
A ClustVarSelect object
}
\description{
Create ClustVarSelect Object
}
\keyword{internal}
