% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_kmeans_bregman_s3method.R
\name{create_powerkmeans}
\alias{create_powerkmeans}
\title{Power K-means Result Object}
\usage{
create_powerkmeans(centers, classes, reduced_dim, params = list())
}
\arguments{
\item{centers}{Matrix of cluster centers}

\item{classes}{Vector of cluster assignments}

\item{reduced_dim}{Dimension-reduced data matrix}

\item{params}{List of parameters used}
}
\value{
A PowerKmeans object
}
\description{
Power K-means Result Object
}
\keyword{internal}
