% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bregman_divergence_funs.R
\name{pairwise_bregman}
\alias{pairwise_bregman}
\title{Compute Pairwise Bregman Divergences}
\usage{
pairwise_bregman(X, Y, divergence = "euclidean")
}
\arguments{
\item{X}{First numeric matrix of data points}

\item{Y}{Second numeric matrix of data points}

\item{divergence}{Character specifying divergence type: "euclidean", "kl",
"itakura-saito", or "logistic"}
}
\value{
Matrix of pairwise Bregman divergences between X and Y
}
\description{
Calculates pairwise Bregman divergences between two sets of points using
specified divergence type
}
\details{
Computes D(x,y) = phi(x) - phi(y) - <nabla phi(y), x-y> where phi is determined
by the divergence type.
}
\examples{
X <- matrix(rnorm(6), nrow=2)
Y <- matrix(rnorm(9), nrow=3)
dist <- pairwise_bregman(X, Y, "euclidean")

}
\seealso{
\code{\link{euclidean_bregman}}, \code{\link{kl_bregman}},
\code{\link{itakura_saito_bregman}}, \code{\link{logistic_bregman}}
}
