% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_kmeans_bregman_s3method.R
\name{plot.PowerKmeans}
\alias{plot.PowerKmeans}
\title{Plot Method for PowerKmeans Objects}
\usage{
\method{plot}{PowerKmeans}(x, ...)
}
\arguments{
\item{x}{A PowerKmeans object}

\item{...}{Additional arguments passed to plot}
}
\value{
Invisibly returns the input object
}
\description{
Plot Method for PowerKmeans Objects
}
