% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD_s3method.R
\name{print.ClustVarSelect}
\alias{print.ClustVarSelect}
\title{Print Method for ClustVarSelect Objects}
\usage{
\method{print}{ClustVarSelect}(x, ...)
}
\arguments{
\item{x}{A ClustVarSelect object returned by SMD function}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for ClustVarSelect Objects
}
