% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_kmeans_bregman_s3method.R
\name{print.PowerKmeans}
\alias{print.PowerKmeans}
\title{Print Method for PowerKmeans Objects}
\usage{
\method{print}{PowerKmeans}(x, ...)
}
\arguments{
\item{x}{A PowerKmeans object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for PowerKmeans Objects
}
