% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD_s3method.R
\name{summary.ClustVarSelect}
\alias{summary.ClustVarSelect}
\title{Summary Method for ClustVarSelect Objects}
\usage{
\method{summary}{ClustVarSelect}(object, ...)
}
\arguments{
\item{object}{A ClustVarSelect object returned by SMD function}

\item{...}{Additional arguments passed to summary}
}
\value{
A list containing summary statistics
}
\description{
Summary Method for ClustVarSelect Objects
}
